feat(common): separate functionality from common select
