+++ packages/common/src/use-select/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './use-select';
+export * from './use-select';
